requirements.txt
PySide6==6.7.2
sounddevice==0.4.6
numpy==1.26.4
faster-whisper==1.2.1
ctranslate2==4.6.3
jinja2==3.1.4
python-Levenshtein==0.25.1
requests==2.32.3
